♻️ refactor: move import to workspace page, keep export in project detail
- Remove import button from ExportImportPanel — panel is now export-only
- Add ImportProjectButton component to workspace-page-content.tsx — self-contained, calls onSuccess to refresh the project list without page reload
- Add onImportProject prop to WorkspacePageContentProps and wire it to loadProjects in WorkspacePage
